Tests (unit tests, E2E, integration tests)

+++ src/main/resources/db/migration/V6__add_tables_project_and_project_steps.sql
@@ -12,7 +12,7 @@ create table project_steps(
                         id int primary key auto_increment,
                         description varchar(100) not null,
                         project_id int not null,
-                        days_to_deadline bigint,
+                        days_to_deadline int,
                         foreign key (project_id) references projects(id)
 );
